requirements.txt
pandas
numpy
pymongo
pymongo[srv]==3.6
python-dotenv
certifi
scikit-learn
##-e .